#5 rename file "domain.py" to "asset.py"
